fix: preset format
