+++ packages/preset/src/shortcuts/alert.ts
@@ -1,0 +1,11 @@
+export const alertShortcuts: Record<string, string> = {
+  // alert
+  'o-alert-base': 'pr p-3  fi rounded o-transition w-full of-hidden c-context shadow-context:50 o-solid border fi',
+  'o-alert-light': 'o-light',
+  'o-alert-content': 'flex flex-col justify-center w-full',
+  'o-alert-header': 'fb items-center',
+  'o-alert-title': 'flex-1 text-sm font-600',
+  'o-alert-close': 'text-md cursor-pointer pa top-3 right-3 fcc',
+  'o-alert-icon': 'fi pr-5',
+  'o-alert-description': 'flex break-all items-center text-sm',
+}
