+++ packages/preset/src/shortcuts/badge.ts
@@ -1,0 +1,7 @@
+export const badgeShortcuts: Record<string, string> = {
+  // Badge
+  'o-badge': 'pr inline-block align-middle',
+  'o-badge-sup': 'bg-context text-xs min-w-4.5 h-4.5 leading-4.5 whitespace-nowrap fcc inline-flex rounded-full px-1 z-1',
+  'o-badge-sup-fixed': 'pa top-0 right-0 translate-x-[calc(50%-1px)] translate-y-[calc(-50%+1px)]',
+  'o-badge-sup-dot': '!min-w-auto !h-2 !w-2 rounded-full',
+}
