+++ packages/preset/src/shortcuts/button.ts
@@ -1,0 +1,15 @@
+export const buttonShortcuts: Record<string, string> = {
+  // button
+  'o-button-base':
+  'fcc gap-1 inline-flex cursor-pointer rounded shadow-context:50 !outline-none c-context o-solid border',
+  'o-button-hover':
+  'hover-border-context:64 hover-text-white hover-bg-context:64',
+  'o-button-active':
+  'active-border-context:32 active-text-white active-bg-context:32',
+  'o-button-light': 'o-light',
+  'o-button-text': 'o-text',
+  'o-button-xs': 'px-2.5 h-6 text-xs shadow-xs',
+  'o-button-sm': 'px-3 h-7 text-sm shadow-sm',
+  'o-button-md': 'px-4 h-8 text-sm shadow-md',
+  'o-button-lg': 'px-4 h-9 text-md shadow-lg',
+}
