+++ packages/preset/src/shortcuts/checkbox.ts
@@ -1,0 +1,9 @@
+export const checkboxShortcuts: Record<string, string> = {
+  // checkbox
+  'o-checkbox-base': 'fsc gap-1 inline-flex cursor-pointer',
+  'o-checkbox': 'o-transition text-md rounded border fcc peer-checked-(bg-context b-context scale-full children-scale-full)',
+  'o-checkbox-icon': 'o-transition scale-0 text-(white md)',
+  'o-checkbox-md': 'w-4 h-4',
+  'o-checkbox-lg': 'w-5 h-5',
+  'o-checkbox-disabled': 'o-disabled',
+}
