+++ packages/preset/src/shortcuts/common.ts
