+++ packages/preset/src/shortcuts/icon.ts
@@ -1,0 +1,4 @@
+export const iconShortcuts: Record<string, string> = {
+  // icon
+  'o-icon-base': 'c-context text-md',
+}
