+++ packages/preset/src/shortcuts/tag.ts
@@ -1,0 +1,14 @@
+export const tagShortcuts: Record<string, string> = {
+  // tag
+  'o-tag-base': 'fcc gap-1 rounded inline-flex box-border !outline-none c-context o-solid  mx-1',
+  'o-tag-is-closable': 'pr-7px',
+  'o-tag-xs': 'px-1.5 h-5 text-xs',
+  'o-tag-sm': 'px-2.5 h-6 text-sm',
+  'o-tag-md': 'px-3 h-7 text-sm ',
+  'o-tag-lg': 'px-4 h-8 text-md',
+  'o-tag-plain': 'bg-context:20',
+  'o-tag-light': 'c-context border !border-context:60 bg-context:12',
+  'o-tag-disabled': 'o-disabled',
+  'o-tag-rounded': 'rounded-full',
+  'o-tag-close': 'ml-1.5 cursor-pointer c-context',
+}
