+++ packages/preset/src/type.ts
@@ -1,0 +1,2 @@
+export type SizeType = 'xs' | 'sm' | 'md' | 'lg'
+export type ThemeType = 'primary' | 'secondary' | 'success' | 'warning' | 'error' | 'info'
