+++ volar.d.ts
@@ -4,6 +4,7 @@ declare module 'vue' {
     OAlert: typeof import('@onu-ui/components')['OAlert']
     OAvatar: typeof import('@onu-ui/components')['OAvatar']
     OAvatarGroup: typeof import('@onu-ui/components')['OAvatarGroup']
+    OBadge: typeof import('@onu-ui/components')['OBadge']
     OButton: typeof import('@onu-ui/components')['OButton']
     OCard: typeof import('@onu-ui/components')['OCard']
     OCheckbox: typeof import('@onu-ui/components')['OCheckbox']
